remove StructWrapper
I was using this when I was experimenting
with NSCache. Objective-C is too slow
so I decided to scrap NSCache and go with
dictionaries. The only obvious problem is
that dictionaries do not maintain their
memory usage, and may end up using a lot
of RAM. I'll have to write some code to
deal with that inevitability at some point